Moved to a SignalCache svelte component
