+++ ui/src/routes/direct-chats/[agentId]/+layout.svelte
@@ -1,14 +1,12 @@
 <script lang="ts">
 	import { page } from '$app/state';
-	import { createSignalCache } from '$lib/stores/signal-cache.svelte';
+	import SignalCache from '$lib/stores/SignalCache.svelte';
 
 	let { children } = $props();
-
-	// Keep direct-chat reactive promises (peerProfile, messageSets, etc.)
-	// cached while the user is anywhere inside /direct-chats/[agentId]/
-	createSignalCache();
 </script>
 
 {#key page.params.agentId}
-	{@render children()}
+	<SignalCache>
+		{@render children()}
+	</SignalCache>
 {/key}
